requirements.txt
alembic==0.9.6
astroid==1.5.3
Babel==2.5.1
bleach==2.1.1
blinker==1.4
click==6.7
decorator==4.1.2
dominate==2.3.1
Flask==0.12.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-Migrate==2.1.1
Flask-PageDown==0.2.2
flask-paginate==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
html5lib==1.0b10
ipython==6.2.1
ipython-genutils==0.2.0
isort==4.2.15
itsdangerous==0.24
jedi==0.11.0
Jinja2==2.9.6
lazy-object-proxy==1.3.1
Mako==1.0.7
Markdown==2.6.9
MarkupSafe==1.0
mccabe==0.6.1
olefile==0.44
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pylint==1.7.4
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.3
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.1.15
traitlets==4.3.2
visitor==0.1.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
wrapt==1.10.11
WTForms==2.1
